S03: GUI hello world
The main program now says hello in a graphical window. The UI is made
by default of two layers: one background and one overlay layer (which
can be used for drag image) which we demonstrate.

HiDPI support on Windows10 is implemented. The program will not work
out of the box on a <Windows10 machine.

Dependencies
- added khronos opengl headers, glad (gl loader)
- added NanoVG (canvas API)

Libraries
- added uumu audio/video platform layer